allow directory backup

+++ backup.py
@@ -3,6 +3,7 @@ import time
 import shutil
 import argparse
 import datetime
+from pathlib import Path
 
 DEFAULT_INTERVAL = 5
 BACKUP_DEFAULT_DIR = "Backups_dir"
@@ -17,29 +18,53 @@ parser.add_argument('-t', '--time', default=5, type=int, metavar="Interval", des
 
 args = parser.parse_args()
 
-if args.file_path is None:
-    raise argparse.ArgumentError('Filepath must not be none')
-file_path: str = args.file_path
-file_parts = file_path.split('\\')[-1].split('.')
-
 if args.interval is None:
     args.interval = DEFAULT_INTERVAL
 interval: int = args.interval
 
-if args.backup_path is None:
-    path_parts = file_path.rsplit('\\',  maxsplit=1)
-    # if only one argument -> only filename was given -> dir = .
-    args.backup_path = f'./{BACKUP_DEFAULT_DIR}' if len(path_parts)==1 else f"{path_parts[0]}/{BACKUP_DEFAULT_DIR}" #save in dir of original file
-backup_path: str = args.backup_path
+if args.file_path is None:
+    raise ValueError('Filepath must not be none')
+file_path: str = args.file_path
 
-file_name = file_parts[0]
-file_ending = file_parts[1]
+if os.path.isfile(file_path):
+    ## backup a file
+    file_parts = file_path.split('\\')[-1].split('.')
 
-os.makedirs(backup_path, exist_ok=True)
+    if args.backup_path is None:
+        path_parts = file_path.rsplit('\\',  maxsplit=1)
+        # if only one argument -> only filename was given -> dir = .
+        args.backup_path = f'./{BACKUP_DEFAULT_DIR}' if len(path_parts)==1 else f"{path_parts[0]}/{BACKUP_DEFAULT_DIR}" #save in dir of original file
+    backup_path: str = args.backup_path
 
-while True:
-    timestring = datetime.datetime.now().strftime('_%d.%m.%y_%H.%M.%S')
-    new_filepath = f"{backup_path}\\{file_name}{timestring}.{file_ending}"
-    shutil.copy2(file_path, new_filepath)
-    print(f"Stored new Backup {new_filepath}")
-    time.sleep(interval*60)
+    file_name = file_parts[0]
+    file_ending = file_parts[1]
+
+    os.makedirs(backup_path, exist_ok=True)
+
+    while True:
+        timestring = datetime.datetime.now().strftime('_%d.%m.%y_%H.%M.%S')
+        new_filepath = f"{backup_path}\\{file_name}{timestring}.{file_ending}"
+        shutil.copy2(file_path, new_filepath)
+        print(f"Stored new Backup {new_filepath}")
+        time.sleep(interval*60)
+
+
+if os.path.isdir(file_path): ## backup a directory
+    path = Path(os.path.abspath(file_path))
+    dir_name = path.name
+    if args.backup_path is None:
+        ## store in parent directory of backup dir
+        parent_dir = path.parents[0]
+        print(parent_dir)
+        print(path.name)
+        # if only one argument -> only filename was given -> dir = .
+        args.backup_path = f"{parent_dir}/{BACKUP_DEFAULT_DIR}" #save in dir of original file
+        #os.makedirs(args.backup_path, exist_ok=True)
+    backup_path: str = args.backup_path
+
+    while True:
+        timestring = datetime.datetime.now().strftime('_%d.%m.%y_%H.%M.%S')
+        new_filepath = f"{backup_path}\\{dir_name}{timestring}"
+        shutil.copytree(file_path, new_filepath)
+        print(f"Stored new Backup {new_filepath}")
+        time.sleep(interval)
